// src/app/shared/models/custom-component-config.ts
export interface ComponentConfiguration{
    isKeyField: boolean;
    dblClick: boolean;
    mouseOver: boolean;
    onBlur: boolean;
    backgroundColor: string;
    borderColor: string;
    borderHighlight: boolean;
    highlightField: boolean;
    name: string;
    subText: string;
    isSeNumber: boolean;
    isCodeDroplist: boolean;
    isDatePicker: boolean;
}